Add new folder and move QCRerunDisabledModal into it
